+++ model-management/train.py
@@ -10,8 +10,8 @@ from datasets import load_dataset
 from peft import LoraConfig, get_peft_model
 
 from mini_llm.data import create_alpaca_prompt, create_alpaca_prompt_with_response
-from mini_llm.utils import freeze, parse_args, LLMSampleCB
-from mini_llm.hf import debug_trainer_data
+from mini_llm.utils import parse_args, 
+from mini_llm.hf import debug_trainer_data, freeze, LLMSampleCB
 
 
 ALPACA_TOTAL_PACKED_SAMPLES = 11_210 # at seq_len=1024
